to_sym on attributes given to MachineCollection
Earlier today we noticed an issue in one of our projects where the state
wouldn't be assigned on initialization when provided as a string (state
attributes are *usually* kept as a symbol by default)

This would fix the issue but probably bring some performance / security
penalties for projects running on older ruby versions

Ultimately this should also highlight the project is in a tough
situation. The `attributes` arg I had added to initialize_states was
to fix issues related to ativemodel / activerecord initialization but it
ends up making it very confusing if you look at this gem alone. I don't
understand most of it but the root idea of adding state machines to any
ruby object and on top of that work with the very specific and
complicated initialization process for both activerecord and activemodel
might be too much. Picking either one or another to support could makes
things simpler