Fix states evaluated with wrong `owner_class`
When a sub-class clones a new copy of the machine the states get updated but the cached `StateContext` was still pointing to the old machine instance. Meaning that each `state` rule gets evaluated in the context of the parent class.

We first noticed this with ActiveRecord objects using single table inheritance. Adding validation rules in `state`s on the sub-class would cause the parent class to also apply those validations.

The added test case actually passes without the fix, but I can’t quite figure out why. Something to do with the _default definee_ for `def` in the context of a `class_eval`, I think.

In any case, `example_class_method` will cause an undefined method error if it is called on `Vehicle`, but executes correctly on `Motorcycle` with the fix applied.