Support prepend module
When `owner_class` has a defined action, if try to find this action with the
`owner_class_ancestor_has_method?` this method should return `nil` if it
is not defined as module in ancestors list.

The method searches through the list of ancestors by excluding the first
element of the list and after filtering out only the classes.

If a module is prepended then excluding the first element
of the list, the prepended module, and then the first class into search
is the owner_class

I reversed the flow so that before the classes are filtered and then is
excluding the first element of the list, it this way the owner_class is
excluded.